add distributed training examples and local training example
