+++ trainer-redesign/distributed-training-mode/processed/data/dataset_dict.json
@@ -1,0 +1,1 @@
+{"splits": ["train", "validation"]}
