+++ trainer-redesign/local-training/requirements.txt
@@ -1,0 +1,3 @@
+numpy
+pandas
+scikit-learn
