+++ trainer-redesign/local-training/train.py
@@ -1,0 +1,38 @@
+from sklearn.datasets import load_iris
+from sklearn.model_selection import train_test_split
+from sklearn.ensemble import RandomForestClassifier
+from sklearn.metrics import accuracy_score
+import argparse
+
+def main(args):
+    n_estimators = args.n_estimators
+    random_state = args.random_state
+    
+    # Load the Iris dataset
+    iris = load_iris()
+    X = iris.data
+    y = iris.target
+
+    # Split the dataset into training and testing sets
+    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=random_state)
+
+    # Train a RandomForestClassifier
+    clf = RandomForestClassifier(n_estimators=n_estimators, random_state=random_state)
+    clf.fit(X_train, y_train)
+    
+    # Make predictions on the test set
+    y_pred = clf.predict(X_test)
+
+    # Evaluate the model
+    accuracy = accuracy_score(y_test, y_pred)
+    print(f"Model accuracy: {accuracy:.2f}")
+    
+    
+if __name__ == "__main__":
+    print("Training model...")
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--n_estimators", type=int, default=100)
+    parser.add_argument("--random_state", type=int, default=42)
+    args = parser.parse_args()
+    main(args)
+
